requirements.txt
# Backend - FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database — SQLite (dev) + PostgreSQL (prod)
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.13.3

# HTTP clients
httpx==0.27.2
aiohttp>=3.9.5
requests>=2.31.0

# LINE Bot
line-bot-sdk==3.11.0

# Backtest
pandas==2.2.3
numpy==2.1.2
matplotlib==3.9.2

# Scraper / Sentiment
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
anthropic==0.40.0

# Utilities
python-dotenv==1.0.1
apscheduler==3.10.4
loguru==0.7.2
pytz==2024.2
python-dateutil==2.9.0
pillow==10.4.0
